+++ changelog.md
@@ -8,6 +8,8 @@
 - Bulk Update Messages API
 - Remove Unread Messages API
 - Delete Messages APIs
+- Add/Remove fields in CDR Config API Response
+- Add field in ConsentResponse & request
 
 ### 3.18.0